requirements.txt
fastapi
uvicorn
streamlit
requests

langchain
langchain-community
langchain-chroma
langchain-openai

chromadb
sentence-transformers

python-dotenv
pydantic